minor refactoring inside the Map class
is_walkable and get_tile_type both cycled over every layer, doing a string
comparison to see if it was the 'special layer'. This is inefficient, and more
importantly obfuscates the implementation of these methods. Instead, have a
special_layer_id that is initialised in the constructor and use that to get
the appropriate layer directly.