Fix linking with wrong GLES implementation
We want to use the broadcom GLES implementation in /opt/vc/lib - not whatever
is under /usr/lib/arm-linux-gnueabihf. The game won't start if it uses the
wrong one.

It seems that some but not all Pis have the libEGL.so and libGLESv2.so files
under the vc directory, but they do (as far as I can tell) all have the
libbrcmEGL.so and libbrcmGLESv2.so equivilants, so we'll link using those
names.